Added Register and login validation
